Create Sum on Subarrays
Here we are taking 2 as the  element to get a positive sum
because in case of n = 2 and k = 2
if we are taking 1 as the element to take positive sum subarrays
[1, -1]  : 0 is not valid and positive-sum subarrays are 1
[1,0] : {0} invalid subarray.
Hence we are taking 2  instead of we can take any positive integer >1 
here we are taking 2 the minimum one, to avoid the summation being greater than 1000
Because then we can make the array negative by adding 1 single element -1000 to stop making positive subarrays. 
[ 0,1,2,3 ..... n]
For each ith element, the number of subarrays it creates is (i+1).
hence we are traversing in a loop,  if k>= (i+1) So we can put a positive integer to this position
and make k as k-(i+1), Else we need to add some negative integer to this such that only k subarrays will be formed.
so -2*(k-1) -1 we can put, this can make starting k element including current element as positive.