Create Sum of Bitwise OR
For each bit position we are calculating how many  1's, 0's are there. Then for each 0'th bit, we will be getting 1 from each element having 1's to the current position.
for all 1's the number of ones we will be getting is (no)(no-1)/2, here "no" denotes "number of ones" 
eg: 1 1 1
->  here we can get 3 pairs having one .
so as we need to return the sum  Hence we will be finding
total no. of one's = no*(no-1)/2 + (no*nz)
sum to be added  =  (1<<currbitpos )*(total no. of ones)
Answer = answer + sum to be added